Strix 2.0.0
Feats
1. AOT兼容: 全面兼容 GraalVM Native 构建.
2. Strix Captcha: 合并至 Strix 主项目.
3. Server Controller: 由于 Native 构建不兼容 JNI, 故移除.
4. Strix OAuth: 由微信授权登录服务重构升级.
5. Strix Pay: 重构支付服务.
6. 依赖更新.

Fixes
1. 修复 Req 中包含 Double 类型导致前后端参数签名不一致的问题.
2. 修复 Stream 语法的 `toList()` 和 `collect(Collectors.toList())` 在 Native 构建时行为不一致的问题.
3. 修复字典同步功能, 解决了 Native 构建后无法通过包扫描找到字典类的问题.
4. 修复扫描不到 controller 上的 `@Anonymous` 注解的问题.

+++ src/main/java/cn/projectan/strix/model/db/PopularityConfig.java
@@ -10,6 +10,7 @@ import lombok.Setter;
 import lombok.experimental.Accessors;
 
 import java.io.Serial;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -56,7 +57,7 @@ public class PopularityConfig extends BaseModel {
     /**
      * 数值倍率
      */
-    private Double magValue;
+    private BigDecimal magValue;
 
     public PopularityConfig(String createBy, String updateBy) {
         super(createBy, updateBy);
